pyproject.toml
[project]
name = "nextcua"
version = "0.1.0"
description = "Model trained to predit next action"
authors = [
    {name = "Realgamer7067",email = "khantjaimin37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "rapidocr (>=3.6.0,<4.0.0)",
    "paddlepaddle-gpu (>=3.0.0,<4.0.0)",
    "opencv-python (>=4.0.0,<5.0.0)",
    "numpy (>=1.24.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
